refactor(code_generator): extract module registry functionality
Extract module registry management from bundler.rs into a dedicated module_registry.rs file. This includes:

- Module cache initialization and population
- Module naming and identifier generation
- Module attribute assignments
- Registry and sys.modules synchronization

Removed unnecessary wrapper methods in bundler.rs and replaced them with direct calls to module_registry functions.

Results:
- Created module_registry.rs with 541 lines (4,754 tokens)
- Reduced bundler.rs from 94,431 to 90,226 tokens (-4,205 tokens)
- All tests passing with 0 clippy warnings
- No changes to functionality or test fixtures
