+++ crates/cribo/src/code_generator/bundler.rs
@@ -7,9 +7,9 @@ use cow_utils::CowUtils;
 use log::debug;
 use ruff_python_ast::{
     Alias, Arguments, AtomicNodeIndex, Decorator, ExceptHandler, Expr, ExprAttribute, ExprCall,
-    ExprContext, ExprList, ExprName, ExprNoneLiteral, ExprStringLiteral, ExprSubscript, Identifier,
-    Keyword, ModModule, Stmt, StmtAssign, StmtClassDef, StmtFunctionDef, StmtImport,
-    StmtImportFrom, StringLiteral, StringLiteralFlags, StringLiteralValue,
+    ExprContext, ExprList, ExprName, ExprNoneLiteral, ExprStringLiteral, Identifier, Keyword,
+    ModModule, Stmt, StmtAssign, StmtClassDef, StmtFunctionDef, StmtImport, StmtImportFrom,
+    StringLiteral, StringLiteralFlags, StringLiteralValue,
     visitor::source_order::SourceOrderVisitor,
 };
 use ruff_text_size::TextRange;
@@ -1454,7 +1454,18 @@ impl<'a> HybridStaticBundler<'a> {
             log::debug!(
                 "Module '{module_name}' was inlined, creating assignments for imported symbols"
             );
-            self.create_assignments_for_inlined_imports(import_from, &module_name, symbol_renames)
+            #[allow(clippy::too_many_arguments)]
+            crate::code_generator::module_registry::create_assignments_for_inlined_imports(
+                import_from,
+                &module_name,
+                symbol_renames,
+                &self.module_registry,
+                &self.inlined_modules,
+                &self.bundled_modules,
+                |local_name, full_module_path| {
+                    self.create_namespace_with_name(local_name, full_module_path)
+                },
+            )
         }
     }
 
@@ -1753,119 +1764,6 @@ impl<'a> HybridStaticBundler<'a> {
         assignments
     }
 
-    /// Create assignments for symbols imported from inlined modules
-    fn create_assignments_for_inlined_imports(
-        &self,
-        import_from: StmtImportFrom,
-        module_name: &str,
-        symbol_renames: &FxIndexMap<String, FxIndexMap<String, String>>,
-    ) -> Vec<Stmt> {
-        let mut assignments = Vec::new();
-
-        for alias in &import_from.names {
-            let imported_name = alias.name.as_str();
-            let local_name = alias.asname.as_ref().unwrap_or(&alias.name);
-
-            // Check if we're importing a module itself (not a symbol from it)
-            // This happens when the imported name refers to a submodule
-            let full_module_path = format!("{module_name}.{imported_name}");
-
-            // Check if this is a module import
-            // First check if it's a wrapped module
-            if self.module_registry.contains_key(&full_module_path) {
-                // For pure static approach, we don't use sys.modules
-                // Instead, we'll handle this as a deferred import
-                log::debug!("Module '{full_module_path}' is a wrapped module, deferring import");
-                // Skip this - it will be handled differently
-                continue;
-            } else if self.inlined_modules.contains(&full_module_path)
-                || self.bundled_modules.contains(&full_module_path)
-            {
-                // Create a namespace object for the inlined module
-                log::debug!(
-                    "Creating namespace object for module '{imported_name}' imported from \
-                     '{module_name}' - module was inlined"
-                );
-
-                // Create a SimpleNamespace-like object with __name__ set
-                let namespace_stmts =
-                    self.create_namespace_with_name(local_name, &full_module_path);
-                assignments.extend(namespace_stmts);
-
-                // Now add all symbols from the inlined module to the namespace
-                // This should come from semantic analysis of what symbols the module exports
-                if let Some(module_renames) = symbol_renames.get(&full_module_path) {
-                    // Add each symbol from the module to the namespace
-                    for (original_name, renamed_name) in module_renames {
-                        // base.original_name = renamed_name
-                        assignments.push(Stmt::Assign(StmtAssign {
-                            node_index: AtomicNodeIndex::dummy(),
-                            targets: vec![Expr::Attribute(ExprAttribute {
-                                node_index: AtomicNodeIndex::dummy(),
-                                value: Box::new(Expr::Name(ExprName {
-                                    node_index: AtomicNodeIndex::dummy(),
-                                    id: local_name.as_str().into(),
-                                    ctx: ExprContext::Load,
-                                    range: TextRange::default(),
-                                })),
-                                attr: Identifier::new(original_name, TextRange::default()),
-                                ctx: ExprContext::Store,
-                                range: TextRange::default(),
-                            })],
-                            value: Box::new(Expr::Name(ExprName {
-                                node_index: AtomicNodeIndex::dummy(),
-                                id: renamed_name.clone().into(),
-                                ctx: ExprContext::Load,
-                                range: TextRange::default(),
-                            })),
-                            range: TextRange::default(),
-                        }));
-                    }
-                }
-            } else {
-                // Regular symbol import
-                // Check if this symbol was renamed during inlining
-                let actual_name = if let Some(module_renames) = symbol_renames.get(module_name) {
-                    module_renames
-                        .get(imported_name)
-                        .map(|s| s.as_str())
-                        .unwrap_or(imported_name)
-                } else {
-                    imported_name
-                };
-
-                // Only create assignment if the names are different
-                if local_name.as_str() != actual_name {
-                    log::debug!(
-                        "Creating assignment: {local_name} = {actual_name} (from inlined module \
-                         '{module_name}')"
-                    );
-
-                    let assignment = StmtAssign {
-                        node_index: AtomicNodeIndex::dummy(),
-                        targets: vec![Expr::Name(ExprName {
-                            node_index: AtomicNodeIndex::dummy(),
-                            id: local_name.as_str().into(),
-                            ctx: ExprContext::Store,
-                            range: TextRange::default(),
-                        })],
-                        value: Box::new(Expr::Name(ExprName {
-                            node_index: AtomicNodeIndex::dummy(),
-                            id: actual_name.into(),
-                            ctx: ExprContext::Load,
-                            range: TextRange::default(),
-                        })),
-                        range: TextRange::default(),
-                    };
-
-                    assignments.push(Stmt::Assign(assignment));
-                }
-            }
-        }
-
-        assignments
-    }
-
     /// Create a namespace object with __name__ attribute
     fn create_namespace_with_name(&self, var_name: &str, module_path: &str) -> Vec<Stmt> {
         // Create: var_name = types.SimpleNamespace()
@@ -2093,14 +1991,6 @@ impl<'a> HybridStaticBundler<'a> {
         } else {
             result_stmts
         }
-    }
-
-    /// Get synthetic module name using content hash
-    fn get_synthetic_module_name(&self, module_name: &str, content_hash: &str) -> String {
-        let module_name_escaped = Self::sanitize_module_name_for_identifier(module_name);
-        // Use first 6 characters of content hash for readability
-        let short_hash = &content_hash[..6];
-        format!("__cribo_{short_hash}_{module_name_escaped}")
     }
 
     /// Check if a string is a valid Python identifier
@@ -2922,7 +2812,7 @@ impl<'a> HybridStaticBundler<'a> {
                                 let alias_is_mandatory = if has_alias {
                                     // Check if there's a local class with the same name as
                                     // import_attr
-                                    self.check_local_name_conflict(ast, &import_attr)
+                                    crate::code_generator::module_registry::check_local_name_conflict(ast, &import_attr)
                                 } else {
                                     false
                                 };
@@ -2948,143 +2838,6 @@ impl<'a> HybridStaticBundler<'a> {
         }
 
         hard_deps
-    }
-
-    /// Generate module cache initialization
-    fn generate_module_cache_init(&mut self) -> Stmt {
-        // __cribo_module_cache__ = {}
-        let assign = StmtAssign {
-            node_index: AtomicNodeIndex::dummy(),
-            targets: vec![Expr::Name(ExprName {
-                node_index: AtomicNodeIndex::dummy(),
-                id: "__cribo_module_cache__".into(),
-                ctx: ExprContext::Store,
-                range: TextRange::default(),
-            })],
-            value: Box::new(Expr::Dict(ruff_python_ast::ExprDict {
-                node_index: AtomicNodeIndex::dummy(),
-                items: vec![],
-                range: TextRange::default(),
-            })),
-            range: TextRange::default(),
-        };
-
-        Stmt::Assign(assign)
-    }
-
-    /// Generate module cache population
-    fn generate_module_cache_population(
-        &mut self,
-        modules: &[(String, ModModule, PathBuf, String)],
-    ) -> Vec<Stmt> {
-        let mut stmts = Vec::new();
-
-        // For each module, add: __cribo_module_cache__["module.name"] = _ModuleNamespace()
-        for (module_name, _, _, _) in modules {
-            let assign = StmtAssign {
-                node_index: AtomicNodeIndex::dummy(),
-                targets: vec![Expr::Subscript(ExprSubscript {
-                    node_index: AtomicNodeIndex::dummy(),
-                    value: Box::new(Expr::Name(ExprName {
-                        node_index: AtomicNodeIndex::dummy(),
-                        id: "__cribo_module_cache__".into(),
-                        ctx: ExprContext::Load,
-                        range: TextRange::default(),
-                    })),
-                    slice: Box::new(Expr::StringLiteral(ExprStringLiteral {
-                        node_index: AtomicNodeIndex::dummy(),
-                        value: StringLiteralValue::single(StringLiteral {
-                            node_index: AtomicNodeIndex::dummy(),
-                            value: module_name.clone().into_boxed_str(),
-                            flags: StringLiteralFlags::empty(),
-                            range: TextRange::default(),
-                        }),
-                        range: TextRange::default(),
-                    })),
-                    ctx: ExprContext::Store,
-                    range: TextRange::default(),
-                })],
-                value: Box::new(Expr::Call(ExprCall {
-                    node_index: AtomicNodeIndex::dummy(),
-                    func: Box::new(Expr::Name(ExprName {
-                        node_index: AtomicNodeIndex::dummy(),
-                        id: "_ModuleNamespace".into(),
-                        ctx: ExprContext::Load,
-                        range: TextRange::default(),
-                    })),
-                    arguments: Arguments {
-                        node_index: AtomicNodeIndex::dummy(),
-                        args: Box::from([]),
-                        keywords: Box::from([]),
-                        range: TextRange::default(),
-                    },
-                    range: TextRange::default(),
-                })),
-                range: TextRange::default(),
-            };
-            stmts.push(Stmt::Assign(assign));
-        }
-
-        stmts
-    }
-
-    /// Generate sys.modules sync
-    fn generate_sys_modules_sync(&mut self) -> Vec<Stmt> {
-        let mut stmts = Vec::new();
-
-        // import sys
-        stmts.push(Stmt::Import(StmtImport {
-            node_index: AtomicNodeIndex::dummy(),
-            names: vec![Alias {
-                node_index: AtomicNodeIndex::dummy(),
-                name: Identifier::new("sys", TextRange::default()),
-                asname: None,
-                range: TextRange::default(),
-            }],
-            range: TextRange::default(),
-        }));
-
-        // sys.modules.update(__cribo_module_cache__)
-        let update_call = Stmt::Expr(ruff_python_ast::StmtExpr {
-            node_index: AtomicNodeIndex::dummy(),
-            value: Box::new(Expr::Call(ExprCall {
-                node_index: AtomicNodeIndex::dummy(),
-                func: Box::new(Expr::Attribute(ExprAttribute {
-                    node_index: AtomicNodeIndex::dummy(),
-                    value: Box::new(Expr::Attribute(ExprAttribute {
-                        node_index: AtomicNodeIndex::dummy(),
-                        value: Box::new(Expr::Name(ExprName {
-                            node_index: AtomicNodeIndex::dummy(),
-                            id: "sys".into(),
-                            ctx: ExprContext::Load,
-                            range: TextRange::default(),
-                        })),
-                        attr: Identifier::new("modules", TextRange::default()),
-                        ctx: ExprContext::Load,
-                        range: TextRange::default(),
-                    })),
-                    attr: Identifier::new("update", TextRange::default()),
-                    ctx: ExprContext::Load,
-                    range: TextRange::default(),
-                })),
-                arguments: Arguments {
-                    node_index: AtomicNodeIndex::dummy(),
-                    args: Box::from([Expr::Name(ExprName {
-                        node_index: AtomicNodeIndex::dummy(),
-                        id: "__cribo_module_cache__".into(),
-                        ctx: ExprContext::Load,
-                        range: TextRange::default(),
-                    })]),
-                    keywords: Box::from([]),
-                    range: TextRange::default(),
-                },
-                range: TextRange::default(),
-            })),
-            range: TextRange::default(),
-        });
-        stmts.push(update_call);
-
-        stmts
     }
 
     /// Process wrapper module globals (matching original implementation)
@@ -3158,160 +2911,6 @@ impl<'a> HybridStaticBundler<'a> {
 
         // Should not happen
         unreachable!("transform_module_to_init_function should return a FunctionDef")
-    }
-
-    /// Generate module init call
-    fn generate_module_init_call(
-        &mut self,
-        synthetic_name: &str,
-        _used_sanitized_names: &mut FxIndexSet<String>,
-    ) -> Vec<Stmt> {
-        let mut statements = Vec::new();
-
-        if let Some(init_func_name) = self.init_functions.get(synthetic_name).cloned() {
-            // Get the original module name for this synthetic name
-            let module_name = self
-                .module_registry
-                .iter()
-                .find(|(_, syn_name)| syn_name == &synthetic_name)
-                .map(|(orig_name, _)| orig_name.to_string())
-                .unwrap_or_else(|| synthetic_name.to_string());
-
-            // Check if this module is a parent namespace that already exists
-            // This happens when a module like 'services.auth' has both:
-            // 1. Its own __init__.py (wrapper module)
-            // 2. Submodules like 'services.auth.manager'
-            let is_parent_namespace = self.module_registry.iter().any(|(name, _)| {
-                name != &module_name && name.starts_with(&format!("{module_name}."))
-            });
-
-            if is_parent_namespace {
-                // For parent namespaces, we need to merge attributes instead of overwriting
-                // Generate code that calls the init function and merges its attributes
-                debug!("Module '{module_name}' is a parent namespace - generating merge code");
-
-                // First, create a variable to hold the init result
-                let init_result_var = "__cribo_init_result";
-                statements.push(Stmt::Assign(StmtAssign {
-                    node_index: AtomicNodeIndex::dummy(),
-                    targets: vec![Expr::Name(ExprName {
-                        node_index: AtomicNodeIndex::dummy(),
-                        id: init_result_var.into(),
-                        ctx: ExprContext::Store,
-                        range: TextRange::default(),
-                    })],
-                    value: Box::new(Expr::Call(ExprCall {
-                        node_index: AtomicNodeIndex::dummy(),
-                        func: Box::new(Expr::Name(ExprName {
-                            node_index: AtomicNodeIndex::dummy(),
-                            id: init_func_name.as_str().into(),
-                            ctx: ExprContext::Load,
-                            range: TextRange::default(),
-                        })),
-                        arguments: Arguments {
-                            node_index: AtomicNodeIndex::dummy(),
-                            args: Box::from([]),
-                            keywords: Box::from([]),
-                            range: TextRange::default(),
-                        },
-                        range: TextRange::default(),
-                    })),
-                    range: TextRange::default(),
-                }));
-
-                // Generate the merge attributes code
-                self.generate_merge_module_attributes(
-                    &mut statements,
-                    &module_name,
-                    init_result_var,
-                );
-
-                // Assign the init result to the module variable
-                statements.push(Stmt::Assign(StmtAssign {
-                    node_index: AtomicNodeIndex::dummy(),
-                    targets: vec![Expr::Name(ExprName {
-                        node_index: AtomicNodeIndex::dummy(),
-                        id: module_name.as_str().into(),
-                        ctx: ExprContext::Store,
-                        range: TextRange::default(),
-                    })],
-                    value: Box::new(Expr::Name(ExprName {
-                        node_index: AtomicNodeIndex::dummy(),
-                        id: init_result_var.into(),
-                        ctx: ExprContext::Load,
-                        range: TextRange::default(),
-                    })),
-                    range: TextRange::default(),
-                }));
-            } else {
-                // Direct assignment for modules that aren't parent namespaces
-                let target_expr = if module_name.contains('.') {
-                    // For dotted modules like models.base, create an attribute expression
-                    let parts: Vec<&str> = module_name.split('.').collect();
-                    let mut expr = Expr::Name(ExprName {
-                        node_index: AtomicNodeIndex::dummy(),
-                        id: parts[0].into(),
-                        ctx: ExprContext::Load,
-                        range: TextRange::default(),
-                    });
-
-                    for (i, part) in parts[1..].iter().enumerate() {
-                        let ctx = if i == parts.len() - 2 {
-                            ExprContext::Store // Last part is Store context
-                        } else {
-                            ExprContext::Load
-                        };
-                        expr = Expr::Attribute(ExprAttribute {
-                            node_index: AtomicNodeIndex::dummy(),
-                            value: Box::new(expr),
-                            attr: Identifier::new(*part, TextRange::default()),
-                            ctx,
-                            range: TextRange::default(),
-                        });
-                    }
-                    expr
-                } else {
-                    // For simple modules, use direct name
-                    Expr::Name(ExprName {
-                        node_index: AtomicNodeIndex::dummy(),
-                        id: module_name.as_str().into(),
-                        ctx: ExprContext::Store,
-                        range: TextRange::default(),
-                    })
-                };
-
-                // Generate: module_name = __cribo_init_synthetic_name()
-                // or: parent.child = __cribo_init_synthetic_name()
-                statements.push(Stmt::Assign(StmtAssign {
-                    node_index: AtomicNodeIndex::dummy(),
-                    targets: vec![target_expr],
-                    value: Box::new(Expr::Call(ExprCall {
-                        node_index: AtomicNodeIndex::dummy(),
-                        func: Box::new(Expr::Name(ExprName {
-                            node_index: AtomicNodeIndex::dummy(),
-                            id: init_func_name.as_str().into(),
-                            ctx: ExprContext::Load,
-                            range: TextRange::default(),
-                        })),
-                        arguments: Arguments {
-                            node_index: AtomicNodeIndex::dummy(),
-                            args: Box::from([]),
-                            keywords: Box::from([]),
-                            range: TextRange::default(),
-                        },
-                        range: TextRange::default(),
-                    })),
-                    range: TextRange::default(),
-                }));
-            }
-        } else {
-            statements.push(Stmt::Pass(ruff_python_ast::StmtPass {
-                node_index: AtomicNodeIndex::dummy(),
-                range: TextRange::default(),
-            }));
-        }
-
-        statements
     }
 
     /// Inline a module
@@ -3637,12 +3236,6 @@ impl<'a> HybridStaticBundler<'a> {
             })),
             range: TextRange::default(),
         })
-    }
-
-    /// Generate registries and hook
-    fn generate_registries_and_hook(&mut self) -> Vec<Stmt> {
-        // No longer needed - we don't use sys.modules or import hooks
-        Vec::new()
     }
 
     /// Sort wrapped modules by dependencies
@@ -4236,7 +3829,9 @@ impl<'a> HybridStaticBundler<'a> {
             });
 
             if !assignment_exists {
-                let reassign = self.create_reassignment(&original, &renamed);
+                let reassign = crate::code_generator::module_registry::create_reassignment(
+                    &original, &renamed,
+                );
                 final_body.push(reassign);
             }
         }
@@ -4670,7 +4265,10 @@ impl<'a> HybridStaticBundler<'a> {
             );
 
             // Register module with synthetic name using content hash
-            let synthetic_name = self.get_synthetic_module_name(module_name, content_hash);
+            let synthetic_name = crate::code_generator::module_registry::get_synthetic_module_name(
+                module_name,
+                content_hash,
+            );
             self.module_registry
                 .insert(module_name.clone(), synthetic_name.clone());
 
@@ -5093,15 +4691,18 @@ impl<'a> HybridStaticBundler<'a> {
             let namespace_class = self.generate_module_namespace_class();
             final_body.push(namespace_class);
 
-            let cache_init = self.generate_module_cache_init();
+            let cache_init = crate::code_generator::module_registry::generate_module_cache_init();
             final_body.push(cache_init);
 
             // Populate cache with all wrapper modules
-            let cache_population = self.generate_module_cache_population(&sorted_wrapper_modules);
+            let cache_population =
+                crate::code_generator::module_registry::generate_module_cache_population(
+                    &sorted_wrapper_modules,
+                );
             final_body.extend(cache_population);
 
             // Add sys.modules sync
-            let sys_sync = self.generate_sys_modules_sync();
+            let sys_sync = crate::code_generator::module_registry::generate_sys_modules_sync();
             final_body.extend(sys_sync);
         }
 
@@ -5168,9 +4769,20 @@ impl<'a> HybridStaticBundler<'a> {
                     // Initialize the wrapper module immediately after defining it
                     // ONLY for non-module-cache mode
                     if !use_module_cache_for_wrappers {
-                        let mut temp_names = FxIndexSet::default();
                         let init_stmts =
-                            self.generate_module_init_call(&synthetic_name, &mut temp_names);
+                            crate::code_generator::module_registry::generate_module_init_call(
+                                &synthetic_name,
+                                module_name,
+                                self.init_functions.get(&synthetic_name).map(|s| s.as_str()),
+                                &self.module_registry,
+                                |statements, module_name, init_result_var| {
+                                    self.generate_merge_module_attributes(
+                                        statements,
+                                        module_name,
+                                        init_result_var,
+                                    );
+                                },
+                            );
                         final_body.extend(init_stmts);
                     }
                     // For module cache mode, initialization happens later in dependency order
@@ -5422,7 +5034,8 @@ impl<'a> HybridStaticBundler<'a> {
             }
 
             // Now add the registries after init functions are defined
-            final_body.extend(self.generate_registries_and_hook());
+            final_body
+                .extend(crate::code_generator::module_registry::generate_registries_and_hook());
         }
 
         // Initialize wrapper modules in dependency order AFTER inlined modules are defined
@@ -6045,12 +5658,30 @@ impl<'a> HybridStaticBundler<'a> {
             }
 
             // Add init calls first
-            let mut entry_used_sanitized_names = FxIndexSet::default();
             for synthetic_name in needed_init_calls {
                 // Note: This is in a context where we can't mutate self, so we'll rely on
                 // the namespaces being pre-created by identify_required_namespaces
-                let init_stmts = self
-                    .generate_module_init_call(&synthetic_name, &mut entry_used_sanitized_names);
+                // Get the original module name for this synthetic name
+                let module_name = self
+                    .module_registry
+                    .iter()
+                    .find(|(_, syn_name)| *syn_name == &synthetic_name)
+                    .map(|(orig_name, _)| orig_name.to_string())
+                    .unwrap_or_else(|| synthetic_name.clone());
+
+                let init_stmts = crate::code_generator::module_registry::generate_module_init_call(
+                    &synthetic_name,
+                    &module_name,
+                    self.init_functions.get(&synthetic_name).map(|s| s.as_str()),
+                    &self.module_registry,
+                    |statements, module_name, init_result_var| {
+                        self.generate_merge_module_attributes(
+                            statements,
+                            module_name,
+                            init_result_var,
+                        );
+                    },
+                );
                 final_body.extend(init_stmts);
             }
 
@@ -6957,32 +6588,6 @@ impl<'a> HybridStaticBundler<'a> {
         }
     }
 
-    /// Create module attribute assignment
-    pub fn create_module_attr_assignment(&self, module_var: &str, attr_name: &str) -> Stmt {
-        Stmt::Assign(StmtAssign {
-            node_index: AtomicNodeIndex::dummy(),
-            targets: vec![Expr::Attribute(ExprAttribute {
-                node_index: AtomicNodeIndex::dummy(),
-                value: Box::new(Expr::Name(ExprName {
-                    node_index: AtomicNodeIndex::dummy(),
-                    id: module_var.into(),
-                    ctx: ExprContext::Load,
-                    range: TextRange::default(),
-                })),
-                attr: Identifier::new(attr_name, TextRange::default()),
-                ctx: ExprContext::Store,
-                range: TextRange::default(),
-            })],
-            value: Box::new(Expr::Name(ExprName {
-                node_index: AtomicNodeIndex::dummy(),
-                id: attr_name.into(),
-                ctx: ExprContext::Load,
-                range: TextRange::default(),
-            })),
-            range: TextRange::default(),
-        })
-    }
-
     /// Check if a symbol should be exported from a module
     pub fn should_export_symbol(&self, symbol_name: &str, module_name: &str) -> bool {
         // Don't export __all__ itself as a module attribute
@@ -7069,26 +6674,6 @@ impl<'a> HybridStaticBundler<'a> {
         false
     }
 
-    /// Create a reassignment statement (original_name = renamed_name)
-    fn create_reassignment(&self, original_name: &str, renamed_name: &str) -> Stmt {
-        Stmt::Assign(StmtAssign {
-            node_index: AtomicNodeIndex::dummy(),
-            targets: vec![Expr::Name(ExprName {
-                node_index: AtomicNodeIndex::dummy(),
-                id: original_name.into(),
-                ctx: ExprContext::Store,
-                range: TextRange::default(),
-            })],
-            value: Box::new(Expr::Name(ExprName {
-                node_index: AtomicNodeIndex::dummy(),
-                id: renamed_name.into(),
-                ctx: ExprContext::Load,
-                range: TextRange::default(),
-            })),
-            range: TextRange::default(),
-        })
-    }
-
     /// Generate module namespace class definition
     fn generate_module_namespace_class(&mut self) -> Stmt {
         // class _ModuleNamespace:
@@ -7107,35 +6692,6 @@ impl<'a> HybridStaticBundler<'a> {
         };
 
         Stmt::ClassDef(class_def)
-    }
-
-    /// Check if a name conflicts with any local definition in the module
-    fn check_local_name_conflict(&self, ast: &ModModule, name: &str) -> bool {
-        for stmt in &ast.body {
-            match stmt {
-                Stmt::ClassDef(class_def) => {
-                    if class_def.name.as_str() == name {
-                        return true;
-                    }
-                }
-                Stmt::FunctionDef(func_def) => {
-                    if func_def.name.as_str() == name {
-                        return true;
-                    }
-                }
-                Stmt::Assign(assign_stmt) => {
-                    for target in &assign_stmt.targets {
-                        if let Expr::Name(name_expr) = target
-                            && name_expr.id.as_str() == name
-                        {
-                            return true;
-                        }
-                    }
-                }
-                _ => {}
-            }
-        }
-        false
     }
 
     /// Create a string literal expression
@@ -7164,44 +6720,6 @@ impl<'a> HybridStaticBundler<'a> {
             matches!(hoisted, Stmt::Import(hoisted_import)
                 if hoisted_import.names.iter().any(|alias| alias.name.as_str() == module_name))
         })
-    }
-
-    /// Generate a unique symbol name to avoid conflicts
-    fn generate_unique_name(
-        &self,
-        base_name: &str,
-        existing_symbols: &FxIndexSet<String>,
-    ) -> String {
-        if !existing_symbols.contains(base_name) {
-            return base_name.to_string();
-        }
-
-        // Try adding numeric suffixes
-        for i in 1..1000 {
-            let candidate = format!("{base_name}_{i}");
-            if !existing_symbols.contains(&candidate) {
-                return candidate;
-            }
-        }
-
-        // Fallback with module prefix
-        format!("__cribo_renamed_{base_name}")
-    }
-
-    /// Sanitize a module name for use in a Python identifier
-    /// This is a simple character replacement - collision handling should be done by the caller
-    fn sanitize_module_name_for_identifier(name: &str) -> String {
-        name.chars()
-            .map(|c| match c {
-                // Replace common invalid characters with descriptive names
-                '-' => '_',
-                '.' => '_',
-                ' ' => '_',
-                // For other non-alphanumeric characters, replace with underscore
-                c if c.is_alphanumeric() || c == '_' => c,
-                _ => '_',
-            })
-            .collect::<String>()
     }
 
     /// Collect unique imports from an import statement
@@ -7857,7 +7375,7 @@ impl<'a> HybridStaticBundler<'a> {
                                          conditional import"
                                     );
                                     result.push(
-                                        self.create_module_attr_assignment("module", local_name),
+                                        crate::code_generator::module_registry::create_module_attr_assignment("module", local_name),
                                     );
                                 }
                             }
@@ -8630,7 +8148,7 @@ impl<'a> HybridStaticBundler<'a> {
 
     /// Get a unique name for a symbol, using the same pattern as generate_unique_name
     fn get_unique_name(&self, base_name: &str, existing_symbols: &FxIndexSet<String>) -> String {
-        self.generate_unique_name(base_name, existing_symbols)
+        crate::code_generator::module_registry::generate_unique_name(base_name, existing_symbols)
     }
 
     /// Rewrite hard dependencies in a module's AST
